move nested beans from util to a not excluded package
IntelliJ powerful refactoring has done it.
