+++ core/src/test/resources/com/opensymphony/xwork2/beans/util/Cat-conversion.properties
