+++ core/src/test/resources/com/opensymphony/xwork2/beans/util/MyBeanAction-conversion.properties
